+++ BookApi/src/main/java/pl/coderslab/webmvc/model/Book.java
@@ -8,7 +8,7 @@ public class Book {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String isbn;
     private String title;
     private String author;
@@ -17,7 +17,7 @@ public class Book {
 
     public Book() {}
 
-    public Book(int id, String isbn, String title, String author, String publisher, String type) {
+    public Book(Long id, String isbn, String title, String author, String publisher, String type) {
         this.id = id;
         this.isbn = isbn;
         this.title = title;
@@ -34,11 +34,11 @@ public class Book {
         this.type = type;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
